# openapi-spec.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/wallet": {
            "get": {
                "operationId": "WalletController_getWallet",
                "summary": "Get wallet balances",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Wallet"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/wallet/transactions": {
            "get": {
                "operationId": "WalletController_getTransactions",
                "summary": "Get transaction history",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Wallet"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/wallet/transactions/{id}": {
            "get": {
                "operationId": "WalletController_getTransaction",
                "summary": "Get transaction details",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Wallet"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
...
        "/api/v1/wallet/purchase": {
            "post": {
                "operationId": "WalletController_initiatePurchase",
                "summary": "Initiate coin purchase",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Wallet"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/wallet/purchase/confirm": {
            "post": {
                "operationId": "WalletController_confirmPurchase",
                "summary": "Confirm purchase after payment",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Wallet"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/wallet/redeem": {
            "post": {
                "operationId": "WalletController_requestRedemption",
                "summary": "Request SC redemption",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Wallet"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/wallet/redemptions": {
            "get": {
                "operationId": "WalletController_getRedemptions",
                "summary": "Get redemption history",
                "parameters": [
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "minimum": 1,
                            "default": 1,
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "minimum": 1,
                            "maximum": 100,
                            "default": 20,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Wallet"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/wallet/bonuses": {
            "get": {
                "operationId": "WalletController_getBonuses",
                "summary": "Get available bonuses",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Wallet"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/wallet/bonuses/{type}/claim": {
            "post": {
                "operationId": "WalletController_claimBonus",
                "summary": "Claim a bonus",
                "parameters": [
                    {
                        "name": "type",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Wallet"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
...
        "/api/v1/vip/status": {
            "get": {
                "operationId": "VipController_getUserVipStatus",
                "summary": "Get user VIP status with tier, XP, and progress",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "VIP"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/vip/xp-history": {
            "get": {
                "operationId": "VipController_getUserXpHistory",
                "summary": "Get user XP earning history",
                "parameters": [
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "minimum": 1,
                            "default": 1,
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "minimum": 1,
                            "maximum": 100,
                            "default": 20,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "VIP"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/vip/benefits": {
            "get": {
                "operationId": "VipController_getUserBenefits",
                "summary": "Get current tier benefits for user",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "VIP"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/vip/cashback/claim": {
            "post": {
                "operationId": "VipController_claimCashback",
                "summary": "Claim available cashback",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "VIP"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
...
        "/api/v1/verification/2fa/send": {
            "post": {
                "operationId": "VerificationController_send2FACode",
                "summary": "Send 2FA verification code",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Verification"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/verification/2fa/verify": {
            "post": {
                "operationId": "VerificationController_verify2FACode",
                "summary": "Verify 2FA code",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Verification"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "WBC 2026 API",
        "description": "WBC 2026 Social Casino Platform API Documentation",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "JWT-auth": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http",
                "name": "JWT",
                "description": "Enter JWT token",
                "in": "header"
            }
        },
...
    }
}